backend/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
distro==1.9.0
Django==5.0.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
exceptiongroup==1.2.1
filelock==3.14.0
firebase-admin==6.5.0
fsspec==2024.5.0
gcloud==0.18.3
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
Jinja2==3.1.4
joblib==1.4.2
jwcrypto==1.5.6
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
openai==1.30.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
python_avatars==1.4.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.40.2
triton==2.3.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
websocket-client==1.8.0
zope.interface==6.3
